requirements.txt
requests
bs4
pypdf2
prettyconf
typer
logzero
pikepdf
reportlab
fpdf2
python-slugify
